Add a new setting that enables dark-mode colouring.
A new option has been added that allows the user to toggle between light
and dark mode. This new option is not enabled by default.

The way in which it works is to recursively replace the foreground,
background and grid paint brushes with some darker colours that I
sampled from Visual Studio. Additionally, a P/Invoke call is made to
have Windows colour the title bar appropriately.

In order for the settings menu to notify the main form of a change, we
now pass it an instance of MainForm within its constructor which is then
called to set the appropriate mode if the user changed the checkbox.

The settings window does not receive any colour adjustments since I did
not deem it worth the effort to recolour a window that the user is
unlikely to be spending a significant amount of time looking at.

The app version in the manifest has also been updated to match what is
currently defined in the UI.

closes #92