Fix incorrect rounding of Module effect percentages.
Due to the JSON having 32-bit IEEE-754 float values, numbers such as
0.04 are encoded as 0.0399999... which subsequently requires adjusting
to the nearest decimal.

Previously, we were multiplying up by 100, but in the case of the above
number, this results in obtaining 3.99999 and then truncating it to 3.

Now, we multiply by 1000 and use the Math.Round feature to round up and
eliminate any decimals, followed by restoring it to a double precision
representation.

fixes #95