Improve the size of recipe/item icons and enable DPI awareness.
The icons used for items and recipes was unreasonably small even on a
1080p monitor - this is certainly compounded into a terrible experience
for higher resolution monitors.

Consequently, the icons have been adjusted in size to make them clearer
to see. To ensure that the regular form controls also render without
blurriness, the manifest and app config has been adjusted to opt-in to
Windows DPI awareness.

As a usability improvement, the NFButton class also greyscales out its
background image when it gets disabled to make it clearer to the user
that the button is currently not clickable.

fixes #63